requirements.txt
streamlit
pytesseract
pdfplumber
pandas
scikit-learn
PyPDF2
easyocr
pdf2image
numpy
Pillow